requirements.txt
numpy
scipy
pandas
matplotlib
pyyaml
opencv-python
torch
torchvision
onnx
onnxruntime
albumentations
mlflow
jinja2
gymnasium
stable-baselines3
networkx
shapely
pymavlink
mavsdk
prometheus-client
paho-mqtt


gymnasium
stable-baselines3
mavsdk
numpy
